backend/requirements.txt
fastapi
uvicorn
httpx
google-generativeai
polyline
pydantic
python-dotenv
